Moved to kebab case formatting announcement bar (#25546)
ref https://linear.app/ghost/issue/NY-722/migrate-ghost-codebase-to-kebab-case-file-naming-convention

* Renamed all files to match to kebab-case format
* Moved tests to match folder structure in other micro-aps
* Added ESLint kebab-case rule
* Updated `README` with the package release process 
* Added to match our working agreement on filenames in Ghost repo
