+++ apps/announcement-bar/src/components/announcement-bar.css
