+++ apps/announcement-bar/src/components/preview.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import {AnnouncementBar} from './AnnouncementBar';
+import {AnnouncementBar} from './announcement-bar';
 
 export function Preview({previewData}) {
     return (
